# fix: Make bulkCreate run in O(N) time. Closes issue #4247.