--- a/lib/dialects/abstract/query-generator.js
+++ b/lib/dialects/abstract/query-generator.js
@@ -222,7 +222,7 @@ const QueryGenerator = {
 
     const query = 'INSERT<%= ignoreDuplicates %> INTO <%= table %> (<%= attributes %>) VALUES <%= tuples %><%= onDuplicateKeyUpdate %><%= returning %>;';
     const tuples = [];
-    const serials = [];
+    const serials = {};
     const allAttributes = [];
     let onDuplicateKeyUpdate = '';
 
@@ -233,14 +233,14 @@ const QueryGenerator = {
         }
 
         if (rawAttributes[key] && rawAttributes[key].autoIncrement === true) {
-          serials.push(key);
+          serials[key] = true;
         }
       });
     }
 
     for (const attrValueHash of attrValueHashes) {
       tuples.push('(' + allAttributes.map(key => {
-        if (this._dialect.supports.bulkDefault && serials.indexOf(key) !== -1) {
+        if (this._dialect.supports.bulkDefault && serials[key] === true) {
           return attrValueHash[key] || 'DEFAULT';
         }
         return this.escape(attrValueHash[key], rawAttributes[key], { context: 'INSERT' });