### Pull Request check-list

- [x] Does `npm run test` or `npm run test-DIALECT` pass with this change (including linting)?
- [x] Does your issue contain a link to existing issue (Closes #[issue]) or a description of the issue you are solving?
- [x] Have you added new tests to prevent regressions? (performance fix, not required)
- [x] Is a documentation update included (if this change modifies existing APIs, or introduces new ones)? (doesn't change existing APIs)
- [x] Have you added an entry under `Future` in the changelog?

### Description of change

In generating the bulk insert query, the serials are stored in an array. This array can potentially be up to size N, which is the number of rows to be created.

When generating the tuples at a later stage, we walk through the records one-by-one - O(N) iterations. However, in a single iteration, we were performing a lookup in the serials array (indexOf costs O(N)),
leading to a O(N^2) runtime and poor performance for large number of records inserted.

This patch converts the serials array into an object map instead, so lookups cost O(1) time.