@yjwong 2017-03-17
Hmm, not sure why the test failed, looks like something unrelated. Also I found out it still doesn't run in O(N) time, and `sequelize-benchmark` doesn't seem to trigger the slow path.