@yjwong 2017-03-20
As tested with sequelize-benchmark@1.1.0:

**PR**

```
▶ DIALECT=postgres SEQ_USER=postgres SEQ_PW=postgres SEQ_DB=postgres sequelize-benchmark
  sequelize:perf Using postgres DIALECT. +0ms
  sequelize:perf Env ready +242ms
  sequelize:perf Starting findAll +6s
  sequelize:perf finding 1500 rows x 2.31 ops/sec ±1.75% (16 runs sampled) (mean 0.433589s) +7s
  sequelize:perf finding(raw) 1500 rows x 2.49 ops/sec ±0.38% (17 runs sampled) (mean 0.402185s) +7s
  sequelize:perf finding(plain) 1500 rows x 2.38 ops/sec ±0.91% (16 runs sampled) (mean 0.420498s) +7s
  sequelize:perf finding 1500 rows with join x 2.19 ops/sec ±1.12% (15 runs sampled) (mean 0.456895s) +7s
  sequelize:perf finding 1500 rows with join (heavy includes) x 2.00 ops/sec ±0.70% (14 runs sampled) (mean 0.499360s) +7s
  sequelize:perf finding 10 rows x 137 ops/sec ±1.42% (79 runs sampled) (mean 0.007275s) +6s
  sequelize:perf finding(raw) 10 rows x 147 ops/sec ±0.52% (78 runs sampled) (mean 0.006801s) +6s
  sequelize:perf finding(plain) 10 rows x 139 ops/sec ±0.71% (72 runs sampled) (mean 0.007211s) +6s
  sequelize:perf finding 10 rows with join x 125 ops/sec ±0.66% (81 runs sampled) (mean 0.008018s) +6s
  sequelize:perf Starting bulkCreate +147ms
  sequelize:perf inserting 10000 rows x 1.06 ops/sec ±2.01% (10 runs sampled) (mean 0.940509s) +10s
  sequelize:perf inserting 1000 rows x 9.28 ops/sec ±3.57% (48 runs sampled) (mean 0.107709s) +6s
  sequelize:perf inserting 10 rows x 98.37 ops/sec ±4.04% (67 runs sampled) (mean 0.010166s) +6s
  sequelize:perf Benchmark finished +0ms

▶ DIALECT=sqlite sequelize-benchmark
  sequelize:perf Using sqlite DIALECT. +0ms
  sequelize:perf Env ready +125ms
  sequelize:perf Starting findAll +6s
  sequelize:perf finding 1500 rows x 22.91 ops/sec ±2.00% (57 runs sampled) (mean 0.043651s) +6s
  sequelize:perf finding(raw) 1500 rows x 51.32 ops/sec ±1.10% (79 runs sampled) (mean 0.019486s) +6s
  sequelize:perf finding(plain) 1500 rows x 22.91 ops/sec ±1.98% (57 runs sampled) (mean 0.043654s) +6s
  sequelize:perf finding 1500 rows with join x 9.81 ops/sec ±2.73% (50 runs sampled) (mean 0.101903s) +6s
  sequelize:perf finding 1500 rows with join (heavy includes) x 7.00 ops/sec ±1.87% (38 runs sampled) (mean 0.142853s) +6s
  sequelize:perf finding 10 rows x 249 ops/sec ±0.84% (81 runs sampled) (mean 0.004018s) +6s
  sequelize:perf finding(raw) 10 rows x 262 ops/sec ±0.99% (85 runs sampled) (mean 0.003811s) +6s
  sequelize:perf finding(plain) 10 rows x 247 ops/sec ±0.96% (81 runs sampled) (mean 0.004041s) +6s
  sequelize:perf finding 10 rows with join x 200 ops/sec ±1.54% (83 runs sampled) (mean 0.005001s) +6s
  sequelize:perf Starting bulkCreate +45ms
  sequelize:perf inserting 10000 rows x 1.35 ops/sec ±1.71% (11 runs sampled) (mean 0.738845s) +8s
  sequelize:perf inserting 1000 rows x 12.77 ops/sec ±1.80% (62 runs sampled) (mean 0.078333s) +6s
  sequelize:perf inserting 10 rows x 500 ops/sec ±2.39% (76 runs sampled) (mean 0.002000s) +6s
  sequelize:perf Benchmark finished +1ms
```

**Master**

```
▶ DIALECT=postgres SEQ_USER=postgres SEQ_PW=postgres SEQ_DB=postgres sequelize-benchmark
  sequelize:perf Using postgres DIALECT. +0ms
  sequelize:perf Env ready +262ms
  sequelize:perf Starting findAll +14s
  sequelize:perf finding 1500 rows x 2.37 ops/sec ±1.26% (16 runs sampled) (mean 0.421138s) +7s
  sequelize:perf finding(raw) 1500 rows x 2.52 ops/sec ±0.40% (17 runs sampled) (mean 0.397584s) +7s
  sequelize:perf finding(plain) 1500 rows x 2.40 ops/sec ±0.55% (16 runs sampled) (mean 0.417305s) +7s
  sequelize:perf finding 1500 rows with join x 2.22 ops/sec ±1.03% (15 runs sampled) (mean 0.449607s) +7s
  sequelize:perf finding 1500 rows with join (heavy includes) x 1.95 ops/sec ±0.42% (14 runs sampled) (mean 0.513636s) +7s
  sequelize:perf finding 10 rows x 133 ops/sec ±1.20% (77 runs sampled) (mean 0.007537s) +6s
  sequelize:perf finding(raw) 10 rows x 144 ops/sec ±0.45% (82 runs sampled) (mean 0.006967s) +6s
  sequelize:perf finding(plain) 10 rows x 135 ops/sec ±0.54% (78 runs sampled) (mean 0.007396s) +6s
  sequelize:perf finding 10 rows with join x 123 ops/sec ±0.60% (80 runs sampled) (mean 0.008159s) +6s
  sequelize:perf Starting bulkCreate +164ms
  sequelize:perf inserting 10000 rows x 0.58 ops/sec ±1.34% (7 runs sampled) (mean 1.715609s) +12s
  sequelize:perf inserting 1000 rows x 8.55 ops/sec ±3.60% (44 runs sampled) (mean 0.116998s) +6s
  sequelize:perf inserting 10 rows x 99.59 ops/sec ±3.73% (67 runs sampled) (mean 0.010041s) +6s
  sequelize:perf Benchmark finished +0ms

▶ DIALECT=sqlite sequelize-benchmark
  sequelize:perf Using sqlite DIALECT. +0ms
  sequelize:perf Env ready +123ms
  sequelize:perf Starting findAll +5s
  sequelize:perf finding 1500 rows x 22.46 ops/sec ±2.21% (56 runs sampled) (mean 0.044514s) +6s
  sequelize:perf finding(raw) 1500 rows x 50.18 ops/sec ±1.59% (78 runs sampled) (mean 0.019930s) +6s
  sequelize:perf finding(plain) 1500 rows x 22.35 ops/sec ±2.72% (56 runs sampled) (mean 0.044733s) +6s
  sequelize:perf finding 1500 rows with join x 10.14 ops/sec ±2.70% (52 runs sampled) (mean 0.098639s) +6s
  sequelize:perf finding 1500 rows with join (heavy includes) x 6.97 ops/sec ±2.29% (37 runs sampled) (mean 0.143483s) +6s
  sequelize:perf finding 10 rows x 243 ops/sec ±1.02% (75 runs sampled) (mean 0.004114s) +6s
  sequelize:perf finding(raw) 10 rows x 270 ops/sec ±1.71% (81 runs sampled) (mean 0.003708s) +6s
  sequelize:perf finding(plain) 10 rows x 242 ops/sec ±1.22% (79 runs sampled) (mean 0.004133s) +6s
  sequelize:perf finding 10 rows with join x 196 ops/sec ±1.11% (81 runs sampled) (mean 0.005109s) +6s
  sequelize:perf Starting bulkCreate +39ms
  sequelize:perf inserting 10000 rows x 1.32 ops/sec ±2.02% (11 runs sampled) (mean 0.757445s) +9s
  sequelize:perf inserting 1000 rows x 12.29 ops/sec ±2.11% (60 runs sampled) (mean 0.081379s) +6s
  sequelize:perf inserting 10 rows x 499 ops/sec ±2.00% (79 runs sampled) (mean 0.002005s) +6s
  sequelize:perf Benchmark finished +0ms
```